fixed: type 'version' of model_ai changed from int to double.

+++ lib/data/models/model_ai.dart
@@ -15,7 +15,7 @@ class ModelAI {
     String name;
     int id;
     String type;
-    int version;
+    double version;
 
     factory ModelAI.fromJson(Map<dynamic, dynamic> json) => ModelAI(
         name: json["name"],
